refactor(chat-runtime): group cross-cutting test helpers under src/testing
